+++ NEWS.md
@@ -1,6 +1,7 @@
 # plavaan 0.0.2
 
-* Minor changes to speed up optimization process; no visible changes at user-interface level.
+* Added `penalized_est_multistart()` and support for custom optimizer starting values via `start` in `penalized_est()`.
+* Minor changes to speed up the optimization process.
 
 # plavaan 0.0.1